requirements.txt
asgiref==3.10.0
certifi==2026.2.25
charset-normalizer==3.4.4
Django==5.2.7
djangorestframework==3.16.1
idna==3.11
mysqlclient==2.2.8
PyMySQL==1.1.2
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.6.3
